api/services/sms_service.py
import logging
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# SMS configuration
SMS_API_KEY = os.getenv("SMS_API_KEY", "")
SMS_API_SECRET = os.getenv("SMS_API_SECRET", "")
SMS_FROM = os.getenv("SMS_FROM", "Notification")

async def send_sms(phone_number: str, message: str) -> bool:
    """
    Send an SMS message.
    Returns True if successful, False otherwise.
    
    In a real implementation, you would use a service like Twilio, Vonage, etc.
    This is a mock implementation that logs the message instead of sending it.
    """
    try:
        if not phone_number:
            logger.error("Phone number is required to send SMS")
            return False
            
        # For development/testing, log the SMS instead of sending it
        if os.getenv("ENVIRONMENT", "development") == "development":
            logger.info(f"[DEV MODE] SMS to: {phone_number}, Message: {message}")
            return True
            
        # In a real implementation, you would use the SMS provider's API
        # Example using Twilio (would require the twilio package):
        # from twilio.rest import Client
        # client = Client(SMS_API_KEY, SMS_API_SECRET)
        # response = client.messages.create(
        #     body=message,
        #     from_=SMS_FROM,
        #     to=phone_number
        # )
        
        logger.info(f"SMS sent to {phone_number}")
        return True
    except Exception as e:
        logger.exception(f"Failed to send SMS: {str(e)}")
        return False