api/requirements.txt
fastapi==0.111.0
uvicorn==0.28.0
pydantic==2.6.4
sqlalchemy==2.0.27
python-multipart==0.0.9
jinja2==3.1.3
aiosmtplib==2.0.2
python-dotenv==1.0.1
email-validator==2.1.0.post1